PR#8490: HIGify extended connection dialog a bit, fix short/int size mismatch bug in the integer field of an option.
[[originally from CVS; cvs2svn created svn r7456]]

+++ client/repodlgs_common.h
@@ -55,10 +55,10 @@ struct options_settable {
   int type;
   unsigned char category;
 
-  short val;
-  short default_val;
-  short min;
-  short max;
+  int val;
+  int default_val;
+  int min;
+  int max;
 
   char *strval;
   char *default_strval;
